Configura fliyway, adiciona tabela de User e corrige tipos de id (long -> Integer)

+++ src/main/java/com/Thiago279/agendamento_medico/entity/Client.java
@@ -1,10 +1,8 @@
 package com.Thiago279.agendamento_medico.entity;
 
-import com.Thiago279.agendamento_medico.entity.Schedule;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
-
 
 import java.util.List;
 
@@ -14,7 +12,7 @@ public class Client {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false, length = 100)
     private String name;
@@ -33,11 +31,11 @@ public class Client {
 
     // Getters and Setters
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
